+++ R/delta.R
@@ -3,6 +3,7 @@ function(data,
          model, 
          type = c("response", "link", "terms"), 
          vcov = vcov(model),
+         weights = NULL,
          eps = 1e-7,
          ...) {
     # take the derivative of each marginal effect from a model with respect to each model coefficient
@@ -21,19 +22,24 @@ function(data,
     # http://www.soderbom.net/lecture10notes.pdf
     # http://stats.stackexchange.com/questions/122066/how-to-use-delta-method-for-standard-errors-of-marginal-effects
     
-    FUN <- .build_grad_fun(data = data, model = model, type = type, ...)
-    #jacobian <- numDeriv::jacobian(FUN, model[["coefficients"]], method = "simple")
+    FUN <- .build_grad_fun(data = data, model = model, type = type, weights = weights, eps = eps, ...)
     jacobian <- jacobian(FUN, coef(model), eps = eps)
     vout <- diag(jacobian %*% vcov %*% t(jacobian))
     return(vout)
 }
 
-.build_grad_fun <- function(data, model, type = "response", eps = 1e-7, ...) {
+.build_grad_fun <- function(data, model, type = "response", weights = NULL, eps = 1e-7, ...) {
     
     # factory function to return prediction holding data constant but varying coefficients
     FUN <- function(coefs) {
         model[["coefficients"]] <- coefs
-        colMeans(marginal_effects(model = model, data = data, type = type, eps = eps, ...), na.rm = TRUE)
+        if (is.null(weights)) {
+            means <- colMeans(marginal_effects(model = model, data = data, type = type, eps = eps, ...), na.rm = TRUE)
+        } else {
+            me_tmp <- marginal_effects(model = model, data = data, type = type, eps = eps, ...)
+            means <- unlist(stats::setNames(lapply(me_tmp, stats::weighted.mean, w = weights, na.rm = TRUE), names(me_tmp)))
+        }
+        means
     }
     return(compiler::cmpfun(FUN))
 }