+++ R/margins_nnet.R
@@ -18,9 +18,6 @@ function(model,
         names(data_list) <- NA_character_
     }
     
-    # warn about weights
-    warn_for_weights(model)
-    
     # calculate marginal effects
     out <- list()
     for (i in seq_along(data_list)) {
@@ -41,6 +38,8 @@ function(model,
               at = if (is.null(at)) at else names(at),
               type = NULL,
               call = if ("call" %in% names(model)) model[["call"]] else NULL,
+              model_class = class(model),
               vce = vce, 
+              weighted = FALSE,
               iterations = NULL)
 }