+++ R/margins_svyglm.R
@@ -1,8 +1,9 @@
 #' @rdname margins
 #' @export
-margins.glm <- 
+margins.svyglm <- 
 function(model, 
          data = find_data(model, parent.frame()), 
+         design,
          variables = NULL,
          at = NULL, 
          type = c("response", "link", "terms"),
@@ -13,9 +14,22 @@ function(model,
          eps = 1e-7,
          ...) {
     
+    # require survey
+    requireNamespace("survey")
+    
     # match.arg()
     type <- match.arg(type)
     vce <- match.arg(vce)
+    
+    # `design` object
+    if (missing(design)) {
+        message("Note: Estimating marginal effects without survey weights. Specify 'design' to adjust for weighting.")
+        wts <- NULL
+    } else if (!inherits(design, "survey.design")) {
+        stop("'design' must be a 'survey.design' object for models of class 'svyglm'")
+    } else {
+        wts <- weights(design)
+    }
     
     # setup data
     data_list <- build_datalist(data, at = at)
@@ -27,18 +41,22 @@ function(model,
     model[["model"]] <- NULL
     attr(model[["terms"]], ".Environment") <- NULL
     
+    # warn about weights
+    #warn_for_weights(model)
+    
     # calculate marginal effects
     out <- list()
     for (i in seq_along(data_list)) {
-        out[[i]] <- build_margins(model = model, 
-                                  data = data_list[[i]], 
+        out[[i]] <- build_margins(model = model,
+                                  data = data_list[[i]],
                                   variables = variables,
-                                  type = type, 
-                                  vcov = vcov, 
-                                  vce = vce, 
-                                  iterations = iterations, 
-                                  unit_ses = unit_ses, 
-                                  eps = eps, 
+                                  type = type,
+                                  vcov = vcov,
+                                  vce = vce,
+                                  iterations = iterations,
+                                  unit_ses = unit_ses,
+                                  weights = wts,
+                                  eps = eps,
                                   ...)
     }
     
@@ -48,8 +66,7 @@ function(model,
               at = if (is.null(at)) at else names(at),
               type = type,
               call = if ("call" %in% names(model)) model[["call"]] else NULL,
-              model_class = class(model),
-              vce = vce, 
-              weighted = FALSE,
+              vce = vce,
+              weighted = if (is.null(wts)) FALSE else TRUE,
               iterations = if (vce == "bootstrap") iterations else NULL)
 }